Copy(move) sycl_iterator.pass.cpp -> sycl_iterator_buf1.pass.cpp + sycl_iterator_buf2.pass.cpp + sycl_iterator_buf3.pass.cpp
