+++ test/general/sycl_iterator_buf2.pass.cpp
@@ -1,5 +1,5 @@
 // -*- C++ -*-
-//===-- sycl_iterator.pass.cpp --------------------------------------------===//
+//===----------------------------------------------------------------------===//
 //
 // Copyright (C) Intel Corporation
 //
@@ -13,100 +13,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "support/test_config.h"
+#include "sycl_iterator.pass.h"
 
-#include _PSTL_TEST_HEADER(execution)
-#include _PSTL_TEST_HEADER(algorithm)
-#include _PSTL_TEST_HEADER(numeric)
-#include _PSTL_TEST_HEADER(memory)
-#include _PSTL_TEST_HEADER(iterator)
-
-#include "support/utils.h"
-#include "oneapi/dpl/pstl/utils.h"
-
-#include <cmath>
-#include <type_traits>
-
-using namespace TestUtils;
-
-//This macro is required for the tests to work correctly in CI with tbb-backend.
 #if TEST_DPCPP_BACKEND_PRESENT
-#include "support/utils_sycl.h"
-
-template <class KernelName1, typename KernelName2>
-class ForComplexName
-{
-};
-
-struct Flip
-{
-    std::int32_t val;
-    Flip(std::int32_t y) : val(y) {}
-    template <typename T>
-    T
-    operator()(const T& x) const
-    {
-        return val - x;
-    }
-};
-struct Plus
-{
-    template <typename T, typename U>
-    T
-    operator()(const T x, const U y) const
-    {
-        return x + y;
-    }
-};
-
-struct Inc
-{
-    template <typename T>
-    void
-    operator()(T& x) const
-    {
-        ++x;
-    }
-};
-
-template <typename T>
-struct Generator_count
-{
-    T def_val;
-    Generator_count(const T& val) : def_val(val) {}
-    T
-    operator()() const
-    {
-        return def_val;
-    }
-    T
-    default_value() const
-    {
-        return def_val;
-    }
-};
-
-// created just to check destroy and destroy_n correctness
-template <typename T>
-struct SyclTypeWrapper
-{
-    T __value;
-
-    explicit SyclTypeWrapper(const T& value = T{4}) : __value(value) {}
-    ~SyclTypeWrapper() { __value = -2; }
-    bool
-    operator==(const SyclTypeWrapper& other) const
-    {
-        return __value == other.__value;
-    }
-};
-
-// this wrapper is needed to take into account not only kernel name,
-// but also other types (for example, iterator's value type)
-template<typename... T>
-struct policy_name_wrapper{};
-
-using namespace oneapi::dpl::execution;
 
 DEFINE_TEST(test_uninitialized_copy)
 {
@@ -127,9 +36,8 @@ DEFINE_TEST(test_uninitialized_copy)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from uninitialized_copy");
     }
@@ -154,9 +62,8 @@ DEFINE_TEST(test_uninitialized_copy_n)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_copy_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from uninitialized_copy_n");
     }
@@ -180,9 +87,8 @@ DEFINE_TEST(test_uninitialized_move)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_move(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from uninitialized_move");
     }
@@ -207,9 +113,8 @@ DEFINE_TEST(test_uninitialized_move_n)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_move_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
                     "wrong effect from uninitialized_move_n");
@@ -231,10 +136,8 @@ DEFINE_TEST(test_uninitialized_fill)
 
         ::std::uninitialized_fill(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2),
                                   value);
+        wait_and_throw(exec);
 
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2),
                                  value),
@@ -256,9 +159,8 @@ DEFINE_TEST(test_uninitialized_fill_n)
         auto value = T1(2);
 
         ::std::uninitialized_fill_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, value + 1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1),
                     "wrong effect from uninitialized_fill_n");
@@ -284,9 +186,8 @@ DEFINE_TEST(test_uninitialized_default_construct)
 
         ::std::uninitialized_default_construct(make_new_policy<new_kernel_name<Policy, 0>>(exec),
                                              first1 + (n / 3), first1 + (n / 2));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), exp_value),
                     "wrong effect from uninitialized_default_construct");
@@ -311,9 +212,8 @@ DEFINE_TEST(test_uninitialized_default_construct_n)
         host_keys.update_data();
 
         ::std::uninitialized_default_construct_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, exp_value),
                     "wrong effect from uninitialized_default_construct_n");
@@ -337,9 +237,8 @@ DEFINE_TEST(test_uninitialized_value_construct)
 
         ::std::uninitialized_value_construct(make_new_policy<new_kernel_name<Policy, 0>>(exec),
                                            first1 + (n / 3), first1 + (n / 2));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), T1{}),
                     "wrong effect from uninitialized_value_construct");
@@ -363,215 +262,11 @@ DEFINE_TEST(test_uninitialized_value_construct_n)
         host_keys.update_data();
 
         ::std::uninitialized_value_construct_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, T1{}),
                     "wrong effect from uninitialized_value_construct_n");
-    }
-};
-
-DEFINE_TEST(test_destroy)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_destroy)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1{ 2 };
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        ::std::destroy(make_new_policy<policy_name_wrapper<new_kernel_name<Policy, 0>, T1>>(exec), first1 + (n / 3),
-                       first1 + (n / 2));
-        if (!::std::is_trivially_destructible<T1>::value)
-            value = T1{-2};
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value),
-                    "wrong effect from destroy");
-    }
-};
-
-DEFINE_TEST(test_destroy_n)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_destroy_n)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1{ 2 };
-
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        ::std::destroy_n(make_new_policy<policy_name_wrapper<new_kernel_name<Policy, 0>, T1>>(exec), first1, n);
-        if(!::std::is_trivially_destructible<T1>::value)
-            value = T1{-2};
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value),
-                    "wrong effect from destroy_n");
-    }
-};
-
-DEFINE_TEST(test_fill)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_fill)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(2);
-
-        ::std::fill(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2), value);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value), "wrong effect from fill");
-    }
-};
-
-DEFINE_TEST(test_fill_n)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_fill_n)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(2);
-
-        ::std::fill_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, value + 1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1), "wrong effect from fill_n");
-    }
-};
-
-DEFINE_TEST(test_generate)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_generate)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(4);
-
-        ::std::generate(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2),
-                      Generator_count<T1>(value));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value),
-                    "wrong effect from generate");
-    }
-};
-
-DEFINE_TEST(test_generate_n)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_generate_n)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(4);
-
-        ::std::generate_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, Generator_count<T1>(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1),
-                    "wrong effect from generate_n");
-    }
-};
-
-DEFINE_TEST(test_for_each)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_for_each)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(6);
-
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        ::std::fill(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value - 1);
-        host_keys.update_data();
-
-        ::std::for_each(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2), Inc());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        // We call due to SYCL 1.2.1: 4.7.2.3.
-        // If the host memory is modified by the host,
-        // or mapped to another buffer or image during the lifetime of this buffer,
-        // then the results are undefined
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value), "wrong effect from for_each");
-    }
-};
-
-DEFINE_TEST(test_for_each_n)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_for_each_n)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(6);
-
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        ::std::for_each_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, Inc());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1),
-                    "wrong effect from for_each_n");
     }
 };
 
@@ -594,9 +289,7 @@ DEFINE_TEST(test_transform_unary)
         update_data(host_keys, host_vals);
 
         ::std::transform(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + n / 2, last1, first2 + n / 2, Flip(7));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
 
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n / 2, value + 1),
@@ -624,64 +317,11 @@ DEFINE_TEST(test_transform_binary)
         host_keys.update_data();
 
         ::std::transform(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first1, first2, Plus());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, T1(6)),
                     "wrong effect from transform_binary");
-    }
-};
-
-DEFINE_TEST(test_replace)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_replace)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(5);
-
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        ::std::replace(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, value, T1(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1),
-                    "wrong effect from replace");
-    }
-};
-
-DEFINE_TEST(test_replace_if)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_replace_if)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(6);
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        ::std::replace_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1,
-                          oneapi::dpl::__internal::__equal_value<T1>(value), T1(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1),
-                    "wrong effect from replace_if");
     }
 };
 
@@ -702,9 +342,8 @@ DEFINE_TEST(test_replace_copy)
         host_keys.update_data();
 
         ::std::replace_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, value, T1(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value + 1),
                     "wrong effect from replace_copy");
@@ -729,9 +368,8 @@ DEFINE_TEST(test_replace_copy_if)
 
         ::std::replace_copy_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2,
                              oneapi::dpl::__internal::__equal_value<T1>(value), T1(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value + 1),
                     "wrong effect from replace_copy_if");
@@ -756,9 +394,8 @@ DEFINE_TEST(test_copy)
         update_data(host_keys, host_vals);
 
         ::std::copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
                     "wrong effect from copy");
@@ -784,9 +421,8 @@ DEFINE_TEST(test_copy_n)
         update_data(host_keys, host_vals);
 
         ::std::copy_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from copy_n");
     }
@@ -811,9 +447,8 @@ DEFINE_TEST(test_move)
         update_data(host_keys, host_vals);
 
         ::std::move(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
                     "wrong effect from move");
@@ -847,9 +482,7 @@ DEFINE_TEST(test_adjacent_difference)
 
         // test with custom functor
         ::std::adjacent_difference(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, __f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
 
         {
             retrieve_data(host_keys, host_vals);
@@ -869,9 +502,7 @@ DEFINE_TEST(test_adjacent_difference)
         host_vals.update_data();
 
         ::std::adjacent_difference(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
 
         retrieve_data(host_keys, host_vals);
 
@@ -883,65 +514,6 @@ DEFINE_TEST(test_adjacent_difference)
             is_correct = is_correct && *(host_first2 + i) == (*(host_first1 + i) - *(host_first1 + i - 1));
 
         EXPECT_TRUE(is_correct, "wrong effect from adjacent_difference #2");
-    }
-};
-
-DEFINE_TEST(test_reduce)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_reduce)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 /* last1 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(2);
-
-        ::std::fill(host_keys.get(), host_keys.get() + n, T1(0));
-        ::std::fill(host_keys.get() + (n / 3), host_keys.get() + (n / 2), value);
-        host_keys.update_data();
-
-        // without initial value
-        auto result1 = ::std::reduce(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result1 == value * (n / 2 - n / 3), "wrong effect from reduce (1)");
-
-        // with initial value
-        auto init = T1(42);
-        auto result2 = ::std::reduce(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2), init);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result2 == init + value * (n / 2 - n / 3), "wrong effect from reduce (2)");
-    }
-};
-
-DEFINE_TEST(test_transform_reduce_unary)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_transform_reduce_unary)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        auto value = T1(1);
-
-        ::std::fill(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        auto result = ::std::transform_reduce(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, T1(42),
-                                            Plus(), ::std::negate<T1>());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == 42 - n, "wrong effect from transform_reduce (unary + binary)");
     }
 };
 
@@ -963,54 +535,9 @@ DEFINE_TEST(test_transform_reduce_binary)
 
         auto result =
             ::std::transform_reduce(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first1, T1(42));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == n + 42, "wrong effect from transform_reduce (2 binary)");
-    }
-};
-
-DEFINE_TEST(test_min_element)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_min_element)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using IteratorValueType = typename ::std::iterator_traits<Iterator>::value_type;
-
-        IteratorValueType fill_value = IteratorValueType{ static_cast<IteratorValueType>(::std::distance(first, last)) };
-
-        ::std::for_each(host_keys.get(), host_keys.get() + n,
-            [&fill_value](IteratorValueType& it) { it = fill_value-- % 10 + 1; });
-
-        ::std::size_t min_dis = n;
-        if (min_dis)
-        {
-            *(host_keys.get() + min_dis / 2) = IteratorValueType{/*min_val*/ 0 };
-            *(host_keys.get() + n - 1) = IteratorValueType{/*2nd min*/ 0 };
-        }
-        host_keys.update_data();
-
-        auto result_min = ::std::min_element(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-
-        auto expected_min = ::std::min_element(host_keys.get(), host_keys.get() + n);
-
-        EXPECT_TRUE(result_min - first == expected_min - host_keys.get(),
-                    "wrong effect from min_element");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: " << *(host_keys.get() + (result_min - first)) << "["
-                    << result_min - first << "], "
-                    << "expected: " << *(host_keys.get() + (expected_min - host_keys.get())) << "["
-                    << expected_min - host_keys.get() << "]" << ::std::endl;
-#    endif
     }
 };
 
@@ -1036,14 +563,13 @@ DEFINE_TEST(test_adjacent_find)
         // check with no adjacent equal elements
         Iterator result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         Iterator expected = last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #1 no elements)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                     << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check with the last adjacent element
         ::std::size_t max_dis = n;
@@ -1054,14 +580,13 @@ DEFINE_TEST(test_adjacent_find)
         }
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         expected = max_dis > 1 ? last - 2 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #2 the last element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check with an adjacent element
         max_dis = n;
@@ -1074,24 +599,24 @@ DEFINE_TEST(test_adjacent_find)
         }
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         expected = max_dis > 1 ? it - 1 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #3 middle element)");
-#    if _ONEDPL_DEBUG_SYCL
+
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
+
         // check with an adjacent element (no predicate)
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #4 middle element (no predicate))");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check with the first adjacent element
         max_dis = n;
@@ -1102,56 +627,13 @@ DEFINE_TEST(test_adjacent_find)
         }
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         expected = max_dis > 1 ? first : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #5 the first element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
-    }
-};
-
-DEFINE_TEST(test_max_element)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_max_element)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using IteratorValueType = typename ::std::iterator_traits<Iterator>::value_type;
-
-        IteratorValueType fill_value = IteratorValueType{0};
-        ::std::for_each(host_keys.get(), host_keys.get() + n,
-                        [&fill_value](IteratorValueType& it) { it = fill_value-- % 10 + 1; });
-
-        ::std::size_t max_dis = n;
-        if (max_dis)
-        {
-            *(host_keys.get() + max_dis / 2) = IteratorValueType{/*max_val*/ 777};
-            *(host_keys.get() + n - 1) = IteratorValueType{/*2nd max*/ 777};
-        }
-        host_keys.update_data();
-
-        auto expected_max_offset = ::std::max_element(host_keys.get(), host_keys.get() + n) - host_keys.get();
-
-        auto result_max_offset = ::std::max_element(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last) - first;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-
-        host_keys.retrieve_data();
-
-        EXPECT_TRUE(result_max_offset == expected_max_offset, "wrong effect from max_element");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: "      << *(host_keys.get() + result_max_offset)   << "[" << result_max_offset   << "], "
-                    << "expected: " << *(host_keys.get() + expected_max_offset) << "[" << expected_max_offset << "]"
-                    << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
     }
 };
 
@@ -1177,14 +659,13 @@ DEFINE_TEST(test_is_sorted_until)
         // check sorted
         Iterator result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         Iterator expected = last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from is_sorted_until (Test #1 sorted sequence)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check unsorted: the last element
         ::std::size_t max_dis = n;
@@ -1195,15 +676,14 @@ DEFINE_TEST(test_is_sorted_until)
         }
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, comp);
         expected = max_dis > 1 ? last - 1 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected,
                     "wrong effect from is_sorted_until (Test #2 unsorted sequence - the last element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check unsorted: the middle element
         max_dis = n;
@@ -1216,27 +696,26 @@ DEFINE_TEST(test_is_sorted_until)
         }
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last, comp);
         expected = it;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected,
                     "wrong effect from is_sorted_until (Test #3 unsorted sequence - the middle element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
+
         // check unsorted: the middle element (no predicate)
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(
             result == expected,
             "wrong effect from is_sorted_until (Test #4 unsorted sequence - the middle element (no predicate))");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check unsorted: the first element
         if (n > 1)
@@ -1246,406 +725,14 @@ DEFINE_TEST(test_is_sorted_until)
         }
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 4>>(exec), first, last, comp);
         expected = n > 1 ? first + 1 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected,
                     "wrong effect from is_sorted_until (Test #5 unsorted sequence - the first element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
-    }
-};
-
-DEFINE_TEST(test_minmax_element)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_minmax_element)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using IteratorValueType = typename ::std::iterator_traits<Iterator>::value_type;
-        auto fill_value = IteratorValueType{ 0 };
-
-        ::std::for_each(host_keys.get(), host_keys.get() + n, [&fill_value](IteratorValueType& it) { it = fill_value++ % 10 + 1; });
-        ::std::size_t dis = n;
-        if (dis > 1)
-        {
-            auto min_it = host_keys.get() + /*min_idx*/ dis / 2 - 1;
-            *(min_it) = IteratorValueType{/*min_val*/ 0 };
-
-            auto max_it = host_keys.get() + /*max_idx*/ dis / 2;
-            *(max_it) = IteratorValueType{/*max_val*/ 777 };
-        }
-        host_keys.update_data();
-
-        auto expected = ::std::minmax_element(host_keys.get(), host_keys.get() + n);
-        auto expected_min = expected.first - host_keys.get();
-        auto expected_max = expected.second - host_keys.get();
-        ::std::pair<Size, Size> expected_offset = { expected_min, expected_max };
-
-        auto result = ::std::minmax_element(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-        auto result_min = result.first - first;
-        auto result_max = result.second - first;
-
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_min == expected_min && result_max == expected_max, "wrong effect from minmax_element");
-        if (!(result_min == expected_min && result_max == expected_max))
-        {
-            host_keys.retrieve_data();
-
-            auto got_min = host_keys.get() + (result.first - first);
-            auto got_max = host_keys.get() + (result.second - first);
-            ::std::cout << "MIN got: " << got_min << "[" << result_min << "], "
-                        << "expected: " << *(host_keys.get() + expected_offset.first) << "[" << expected_min << "]" << ::std::endl;
-            ::std::cout << "MAX got: " << got_max << "[" << result_max << "], "
-                        << "expected: " << *(host_keys.get() + expected_offset.second) << "[" << expected_max << "]" << ::std::endl;
-        }
-    }
-};
-
-DEFINE_TEST(test_is_sorted)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_is_sorted)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using ValueType = typename ::std::iterator_traits<Iterator>::value_type;
-
-        auto comp = ::std::less<ValueType>{};
-
-        ValueType fill_value{ 0 };
-        ::std::for_each(host_keys.get(), host_keys.get() + n,
-                        [&fill_value](ValueType& value) { value = ++fill_value; });
-        host_keys.update_data();
-
-        // check sorted
-        bool result_bool = ::std::is_sorted(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
-        bool expected_bool = true;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool, "wrong effect from is_sorted (Test #1 sorted sequence)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: " << result_bool << ", "
-                    << "expected: " << expected_bool << ::std::endl;
-#    endif
-
-        // check unsorted: the last element
-        ::std::size_t max_dis = n;
-        if (max_dis > 1)
-        {
-            *(host_keys.get() + n - 1) = ValueType{0};
-            host_keys.update_data();
-        }
-        result_bool = ::std::is_sorted(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, comp);
-        expected_bool = max_dis > 1 ? false : true;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool,
-                    "wrong effect from is_sorted (Test #2 unsorted sequence - the last element)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: " << result_bool << ", "
-                    << "expected: " << expected_bool << ::std::endl;
-#    endif
-
-        // check unsorted: the middle element
-        max_dis = n;
-        if (max_dis > 1)
-        {
-            *(host_keys.get() + max_dis / 2) = ValueType{0};
-            host_keys.update_data();
-        }
-        result_bool = ::std::is_sorted(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last, comp);
-        expected_bool = max_dis > 1 ? false : true;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool,
-                    "wrong effect from is_sorted (Test #3 unsorted sequence - the middle element)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: " << result_bool << ", "
-                    << "expected: " << expected_bool << ::std::endl;
-#    endif
-        // check unsorted: the middle element (no predicate)
-        result_bool = ::std::is_sorted(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool,
-                    "wrong effect from is_sorted (Test #4 unsorted sequence - the middle element (no predicate))");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: " << result_bool << ", "
-                    << "expected: " << expected_bool << ::std::endl;
-#    endif
-
-        // check unsorted: the first element
-        max_dis = n;
-        if (max_dis > 1)
-        {
-            *(host_keys.get() + 1) = ValueType{0};
-            host_keys.update_data();
-        }
-        result_bool = ::std::is_sorted(make_new_policy<new_kernel_name<Policy, 4>>(exec), first, last, comp);
-        expected_bool = max_dis > 1 ? false : true;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool,
-                    "wrong effect from is_sorted Test #5 unsorted sequence - the first element");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got: " << result_bool << ", "
-                    << "expected: " << expected_bool << ::std::endl;
-#    endif
-    }
-};
-
-DEFINE_TEST(test_count)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_count)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using ValueType = typename ::std::iterator_traits<Iterator>::value_type;
-        using ReturnType = typename ::std::iterator_traits<Iterator>::difference_type;
-
-        ValueType fill_value{0};
-        ::std::for_each(host_keys.get(), host_keys.get() + n, [&fill_value](ValueType& value) { value = fill_value++ % 10; });
-        host_keys.update_data();
-
-        // check when arbitrary should be counted
-        ReturnType expected = (n - 1) / 10 + 1;
-        ReturnType result = ::std::count(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, ValueType{0});
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == expected, "wrong effect from count (Test #1 arbitrary to count)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got " << result << ", expected " << expected << ::std::endl;
-#    endif
-
-        // check when none should be counted
-        expected = 0;
-        result = ::std::count(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, ValueType{12});
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == expected, "wrong effect from count (Test #2 none to count)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got " << result << ", expected " << expected << ::std::endl;
-#    endif
-
-        // check when all should be counted
-        ::std::fill(host_keys.get(), host_keys.get() + n, ValueType{7});
-        host_keys.update_data();
-
-        expected = n;
-        result = ::std::count(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, ValueType{7});
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == expected, "wrong effect from count (Test #3 all to count)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got " << result << ", expected " << expected << ::std::endl;
-#    endif
-    }
-};
-
-DEFINE_TEST(test_count_if)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_count_if)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using ValueType = typename ::std::iterator_traits<Iterator>::value_type;
-        using ReturnType = typename ::std::iterator_traits<Iterator>::difference_type;
-
-        ValueType fill_value{0};
-        ::std::for_each(host_keys.get(), host_keys.get() + n, [&fill_value](ValueType& value) { value = fill_value++ % 10; });
-        host_keys.update_data();
-
-        // check when arbitrary should be counted
-        ReturnType expected = (n - 1) / 10 + 1;
-        ReturnType result = ::std::count_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last,
-                                            [](ValueType const& value) { return value % 10 == 0; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == expected, "wrong effect from count_if (Test #1 arbitrary to count)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got " << result << ", expected " << expected << ::std::endl;
-#    endif
-
-        // check when none should be counted
-        expected = 0;
-        result = ::std::count_if(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last,
-                                 [](ValueType const& value) { return value > 10; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == expected, "wrong effect from count_if (Test #2 none to count)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got " << result << ", expected " << expected << ::std::endl;
-#    endif
-
-        // check when all should be counted
-        expected = n;
-        result = ::std::count_if(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last,
-                                 [](ValueType const& value) { return value < 10; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result == expected, "wrong effect from count_if (Test #3 all to count)");
-#    if _ONEDPL_DEBUG_SYCL
-        ::std::cout << "got " << result << ", expected " << expected << ::std::endl;
-#    endif
-    }
-};
-
-DEFINE_TEST(test_is_partitioned)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_is_partitioned)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using ValueType = typename ::std::iterator_traits<Iterator>::value_type;
-
-        if (n < 2)
-            return;
-
-        auto less_than = [](const ValueType& value) -> bool { return value < 10; };
-        auto is_odd = [](const ValueType& value) -> bool { return value % 2; };
-
-        bool expected_bool_less_then = false;
-        bool expected_bool_is_odd = false;
-
-        ValueType fill_value{0};
-        ::std::for_each(host_keys.get(), host_keys.get() + n, [&fill_value](ValueType& value) { value = ++fill_value; });
-        expected_bool_less_then = ::std::is_partitioned(host_keys.get(), host_keys.get() + n, less_than);
-        expected_bool_is_odd = ::std::is_partitioned(host_keys.get(), host_keys.get() + n, is_odd);
-        host_keys.update_data();
-
-        // check sorted
-        bool result_bool = ::std::is_partitioned(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, less_than);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool_less_then, "wrong effect from is_partitioned (Test #1 less than)");
-
-        result_bool = ::std::is_partitioned(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, is_odd);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool_is_odd, "wrong effect from is_partitioned (Test #2 is odd)");
-
-        // The code as below was added to prevent accessor destruction working with host memory
-        ::std::partition(host_keys.get(), host_keys.get() + n, is_odd);
-        expected_bool_is_odd = ::std::is_partitioned(host_keys.get(), host_keys.get() + n, is_odd);
-        host_keys.update_data();
-
-        result_bool = ::std::is_partitioned(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last, is_odd);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result_bool == expected_bool_is_odd, "wrong effect from is_partitioned (Test #3 is odd after partition)");
-    }
-};
-
-DEFINE_TEST(test_any_all_none_of)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_any_all_none_of)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-
-        ::std::iota(host_keys.get(), host_keys.get() + n, T1(0));
-        host_keys.update_data();
-
-        // empty sequence case
-        if (n == 1)
-        {
-            auto res0 = ::std::any_of(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1,
-                                      [n](T1 x) { return x == n - 1; });
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
-            EXPECT_TRUE(!res0, "wrong effect from any_of_0");
-            res0 = ::std::none_of(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, first1,
-                                  [](T1 x) { return x == -1; });
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
-            EXPECT_TRUE(res0, "wrong effect from none_of_0");
-            res0 = ::std::all_of(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, first1,
-                                 [](T1 x) { return x % 2 == 0; });
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
-            EXPECT_TRUE(res0, "wrong effect from all_of_0");
-        }
-        // any_of
-        auto res1 = ::std::any_of(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1,
-                                  [n](T1 x) { return x == n - 1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(res1, "wrong effect from any_of_1");
-        auto res2 = ::std::any_of(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1,
-                                  [](T1 x) { return x == -1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(!res2, "wrong effect from any_of_2");
-        auto res3 = ::std::any_of(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1,
-                                  [](T1 x) { return x % 2 == 0; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(res3, "wrong effect from any_of_3");
-
-        //none_of
-        auto res4 = ::std::none_of(make_new_policy<new_kernel_name<Policy, 6>>(exec), first1, last1,
-                                   [](T1 x) { return x == -1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(res4, "wrong effect from none_of");
-
-        //all_of
-        auto res5 = ::std::all_of(make_new_policy<new_kernel_name<Policy, 7>>(exec), first1, last1,
-                                  [](T1 x) { return x % 2 == 0; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(n == 1 || !res5, "wrong effect from all_of");
+#endif // _ONEDPL_DEBUG_SYCL
     }
 };
 
@@ -1674,15 +761,13 @@ DEFINE_TEST(test_equal)
         auto expected  = new_end - new_start > 0;
         auto result = ::std::equal(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + new_start,
                                    first1 + new_end, first2 + new_start);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(expected == result, "wrong effect from equal with 3 iterators");
         result = ::std::equal(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1 + new_start, first1 + new_end,
                               first2 + new_start, first2 + new_end);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(expected == result, "wrong effect from equal with 4 iterators");
     }
 };
@@ -1707,43 +792,37 @@ DEFINE_TEST(test_find_if)
         {
             auto res0 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1,
                                        [n](T1 x) { return x == n - 1; });
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from find_if_0");
             res0 = ::std::find(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, first1, T1(1));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from find_0");
         }
         // find_if
         auto res1 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1,
                                    [n](T1 x) { return x == n - 1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE((res1 - first1) == n - 1, "wrong effect from find_if_1");
 
         auto res2 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1,
                                    [](T1 x) { return x == -1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res2 == last1, "wrong effect from find_if_2");
 
         auto res3 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1,
                                    [](T1 x) { return x % 2 == 0; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res3 == first1, "wrong effect from find_if_3");
 
         //find
         auto res4 = ::std::find(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, T1(-1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res4 == last1, "wrong effect from find");
     }
 };
@@ -1772,18 +851,16 @@ DEFINE_TEST(test_find_first_of)
 
             auto res =
                 ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1, "Wrong effect from find_first_of_1");
         }
         else if (n >= 2 && n < 10)
         {
             auto res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1,
                                             first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_first_of_2");
 
             // No matches
@@ -1791,9 +868,8 @@ DEFINE_TEST(test_find_first_of)
             host_vals.update_data();
 
             res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_first_of_3");
         }
         else if (n >= 10)
@@ -1809,9 +885,8 @@ DEFINE_TEST(test_find_first_of)
             host_keys.update_data();
 
             auto res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1 + pos2, "Wrong effect from find_first_of_4");
 
             // Add second match
@@ -1819,9 +894,8 @@ DEFINE_TEST(test_find_first_of)
             host_keys.update_data();
 
             res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1 + pos1, "Wrong effect from find_first_of_5");
         }
     }
@@ -1848,14 +922,12 @@ DEFINE_TEST(test_search)
         if (n == 1)
         {
             auto res0 = ::std::search(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from search_00");
             res0 = ::std::search(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from search_01");
         }
         auto res1 = ::std::search(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, last2);
@@ -1864,16 +936,14 @@ DEFINE_TEST(test_search)
         {
             // first n-10 elements of the subsequence are at the beginning of first sequence
             auto res2 = ::std::search(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2 + 10, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res2 - first1 == 5, "wrong effect from search_2");
         }
         // subsequence consists of one element (last one)
         auto res3 = ::std::search(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, last1 - 1, last1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(last1 - res3 == 1, "wrong effect from search_3");
 
         // first sequence contains 2 almost similar parts
@@ -1883,9 +953,8 @@ DEFINE_TEST(test_search)
             host_keys.update_data();
 
             auto res4 = ::std::search(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, first2 + 5, first2 + 6);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res4 == first1, "wrong effect from search_4");
         }
     }
@@ -1913,9 +982,8 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, n - start, T(11));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start, "wrong effect from search_1");
         }
         // Search for sequence in the middle
@@ -1927,17 +995,15 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, end - start, T(22));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start, "wrong effect from search_20");
 
             // Search for sequence of lesser size
             res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last,
                                 ::std::max(end - start - 1, (size_t)1), T(22));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start, "wrong effect from search_21");
         }
         // Search for sequence at the beginning
@@ -1948,9 +1014,8 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last, end, T(33));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first, "wrong effect from search_3");
         }
         // Search for sequence that covers the whole range
@@ -1959,33 +1024,29 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 4>>(exec), first, last, n, T(44));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first, "wrong effect from search_4");
         }
         // Search for sequence which is not there
         {
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 5>>(exec), first, last, 2, T(55));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last, "wrong effect from search_50");
 
             // Sequence is there but of lesser size(see search_n_3)
             res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 6>>(exec), first, last, (n / 3 + 1), T(33));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last, "wrong effect from search_51");
         }
 
         // empty sequence case
         {
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 7>>(exec), first, first, 1, T(5 + n - 1));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first, "wrong effect from search_6");
         }
         // 2 distinct sequences, must find the first one
@@ -2003,9 +1064,8 @@ DEFINE_TEST(test_search_n)
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 8>>(exec), first, last,
                                      ::std::min(end1 - start1, end2 - start2), T(66));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start1, "wrong effect from search_7");
         }
 
@@ -2018,9 +1078,8 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 9>>(exec), first + 1, last, seq_len, T(77));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last, "wrong effect from search_8");
         }
     }
@@ -2047,14 +1106,12 @@ DEFINE_TEST(test_mismatch)
         if (n == 1)
         {
             auto res0 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0.first == first1 && res0.second == first2, "wrong effect from mismatch_00");
             res0 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0.first == first1 && res0.second == first2, "wrong effect from mismatch_01");
         }
         auto res1 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
@@ -2063,9 +1120,8 @@ DEFINE_TEST(test_mismatch)
         {
             // first n-10 elements of the subsequence are at the beginning of first sequence
             auto res2 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2 + 5, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res2.first == last1 - 5 && res2.second == last2, "wrong effect from mismatch_2");
         }
     }
@@ -2091,9 +1147,8 @@ DEFINE_TEST(test_transform_inclusive_scan)
         auto res1 = ::std::transform_inclusive_scan(
             make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, ::std::plus<T1>(),
             [](T1 x) { return x * 2; }, value);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last2, "wrong result from transform_inclusive_scan_1");
 
         retrieve_data(host_keys, host_vals);
@@ -2150,9 +1205,8 @@ DEFINE_TEST(test_transform_exclusive_scan)
         auto res1 =
             ::std::transform_exclusive_scan(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2,
                                           T1{}, ::std::plus<T1>(), [](T1 x) { return x * 2; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last2, "wrong result from transform_exclusive_scan");
 
         auto ii = T1(0);
@@ -2188,9 +1242,8 @@ DEFINE_TEST(test_copy_if)
 
         auto res1 = ::std::copy_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2,
                                    [](T1 x) { return x > -1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last2, "wrong result from copy_if_1");
 
         host_vals.retrieve_data();
@@ -2207,9 +1260,8 @@ DEFINE_TEST(test_copy_if)
 
         auto res2 = ::std::copy_if(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2,
                                  [](T1 x) { return x % 2 == 1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res2 == first2 + (last2 - first2) / 2, "wrong result from copy_if_2");
 
         host_vals.retrieve_data();
@@ -2242,9 +1294,8 @@ DEFINE_TEST(test_remove)
 
         auto pos = (last - first) / 2;
         auto res1 = ::std::remove(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, T1(222 + pos));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last - 1, "wrong result from remove");
 
         host_keys.retrieve_data();
@@ -2279,9 +1330,8 @@ DEFINE_TEST(test_remove_if)
         auto pos = (last - first) / 2;
         auto res1 = ::std::remove_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last,
                                    [=](T1 x) { return x == T1(222 + pos); });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last - 1, "wrong result from remove_if");
 
         host_keys.retrieve_data();
@@ -2324,20 +1374,19 @@ DEFINE_TEST(test_unique_copy)
         auto result_first = first2;
         auto result_last =
             ::std::unique_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, result_first, f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         auto result_size = result_last - result_first;
 
         std::int64_t expected_size = (n - 1) / 4 + 1;
 
         // check
         bool is_correct = result_size == expected_size;
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         if (!is_correct)
             ::std::cout << "buffer size: got " << result_last - result_first << ", expected " << expected_size
                       << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         host_vals.retrieve_data();
         auto host_first2 = host_vals.get();
@@ -2374,19 +1423,18 @@ DEFINE_TEST(test_unique)
         // invoke
         auto f = [](IteratorValueType a, IteratorValueType b) { return a == b; };
         auto result_last = ::std::unique(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         auto result_size = result_last - first;
 
         std::int64_t expected_size = (n - 1) / 4 + 1;
 
         // check
         bool is_correct = result_size == expected_size;
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         if (!is_correct)
             ::std::cout << "buffer size: got " << result_last - first << ", expected " << expected_size << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         host_keys.retrieve_data();
         auto host_first1 = host_keys.get();
@@ -2395,139 +1443,13 @@ DEFINE_TEST(test_unique)
             if (*(host_first1 + i) != i + 1)
             {
                 is_correct = false;
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
                 ::std::cout << "got: " << *(host_first1 + i) << "[" << i << "], "
                           << "expected: " << i + 1 << "[" << i << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
             }
             EXPECT_TRUE(is_correct, "wrong effect from unique");
         }
-    }
-};
-
-DEFINE_TEST(test_partition_copy)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_partition_copy)
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 /* last2 */, Iterator3 first3,
-               Iterator3 /* last3 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, n);
-        TestDataTransfer<UDTKind::eRes,  Size> host_res (*this, n);
-
-        using Iterator1ValueType = typename ::std::iterator_traits<Iterator1>::value_type;
-        using Iterator2ValueType = typename ::std::iterator_traits<Iterator2>::value_type;
-        using Iterator3ValueType = typename ::std::iterator_traits<Iterator3>::value_type;
-        auto f = [](Iterator1ValueType value) { return (value % 3 == 0) && (value % 2 == 0); };
-
-        // init
-        ::std::iota(host_keys.get(), host_keys.get() + n, Iterator1ValueType{0});
-        ::std::fill(host_vals.get(), host_vals.get() + n, Iterator2ValueType{-1});
-        ::std::fill(host_res.get(),   host_res.get() + n, Iterator3ValueType{-2});
-        update_data(host_keys, host_vals, host_res);
-
-        // invoke
-        auto res =
-            ::std::partition_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, first3, f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        retrieve_data(host_keys, host_vals, host_res);
-
-        // init for expected
-        ::std::vector<Iterator2ValueType> exp_true(n, -1);
-        ::std::vector<Iterator3ValueType> exp_false(n, -2);
-        auto exp_true_first = exp_true.begin();
-        auto exp_false_first = exp_false.begin();
-
-        // invoke for expected
-        auto exp = ::std::partition_copy(host_keys.get(), host_keys.get() + n, exp_true_first, exp_false_first, f);
-
-        // check
-        bool is_correct = (exp.first - exp_true_first) == (res.first - first2) &&
-                          (exp.second - exp_false_first) == (res.second - first3);
-#    if _ONEDPL_DEBUG_SYCL
-        if (!is_correct)
-            ::std::cout << "N =" << n << ::std::endl
-                      << "buffer size: got {" << res.first - first2 << "," << res.second - first3 << "}, expected {"
-                      << exp.first - exp_true_first << "," << exp.second - exp_false_first << "}" << ::std::endl;
-#    endif
-
-        for (int i = 0; i < ::std::min(exp.first - exp_true_first, res.first - first2) && is_correct; ++i)
-        {
-            if (*(exp_true_first + i) != *(host_vals.get() + i))
-            {
-                is_correct = false;
-#    if _ONEDPL_DEBUG_SYCL
-                ::std::cout << "TRUE> got: " << *(host_vals.get() + i) << "[" << i << "], "
-                          << "expected: " << *(exp_true_first + i) << "[" << i << "]" << ::std::endl;
-#    endif
-            }
-        }
-
-        for (int i = 0; i < ::std::min(exp.second - exp_false_first, res.second - first3) && is_correct; ++i)
-        {
-            if (*(exp_false_first + i) != *(host_res.get() + i))
-            {
-                is_correct = false;
-#    if _ONEDPL_DEBUG_SYCL
-                ::std::cout << "FALSE> got: " << *(host_res.get() + i) << "[" << i << "], "
-                          << "expected: " << *(exp_false_first + i) << "[" << i << "]" << ::std::endl;
-#    endif
-            }
-        }
-
-        EXPECT_TRUE(is_correct, "wrong effect from partition_copy");
-    }
-};
-
-DEFINE_TEST(test_partition)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_partition)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using IteratorValueType = typename ::std::iterator_traits<Iterator>::value_type;
-
-        // init
-        ::std::iota(host_keys.get(), host_keys.get() + n, IteratorValueType{ 0 });
-        host_keys.update_data();
-
-        // invoke partition
-        auto unary_op = [](IteratorValueType value) { return (value % 3 == 0) && (value % 2 == 0); };
-        auto res = ::std::partition(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, unary_op);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-
-        // check
-        host_keys.retrieve_data();
-        EXPECT_TRUE(::std::all_of(host_keys.get(), host_keys.get() + (res - first), unary_op) &&
-                        !::std::any_of(host_keys.get() + (res - first), host_keys.get() + n, unary_op),
-                    "wrong effect from partition");
-        // init
-        ::std::iota(host_keys.get(), host_keys.get() + n, IteratorValueType{0});
-        host_keys.update_data();
-
-        // invoke stable_partition
-        res = ::std::stable_partition(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, unary_op);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-
-        host_keys.retrieve_data();
-        EXPECT_TRUE(::std::all_of(host_keys.get(), host_keys.get() + (res - first), unary_op) &&
-                        !::std::any_of(host_keys.get() + (res - first), host_keys.get() + n, unary_op) &&
-                        ::std::is_sorted(host_keys.get(), host_keys.get() + (res - first)) &&
-                        ::std::is_sorted(host_keys.get() + (res - first), host_keys.get() + n),
-                    "wrong effect from stable_partition");
     }
 };
 
@@ -2548,9 +1470,8 @@ DEFINE_TEST(test_is_heap_until)
         host_keys.update_data();
 
         auto actual = ::std::is_heap_until(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         // first element is always a heap
         EXPECT_TRUE(actual == first + 1, "wrong result of is_heap_until_1");
 
@@ -2562,9 +1483,8 @@ DEFINE_TEST(test_is_heap_until)
         host_keys.update_data();
 
         actual = ::std::is_heap_until(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(actual == (first + n / 2), "wrong result of is_heap_until_2");
 
         host_keys.retrieve_data();
@@ -2572,211 +1492,9 @@ DEFINE_TEST(test_is_heap_until)
         host_keys.update_data();
 
         actual = ::std::is_heap_until(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(actual == last, "wrong result of is_heap_until_3");
-    }
-};
-
-DEFINE_TEST(test_is_heap)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_is_heap)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        using ValueType = typename ::std::iterator_traits<Iterator>::value_type;
-
-        ::std::iota(host_keys.get(), host_keys.get() + n, ValueType(0));
-        ::std::make_heap(host_keys.get(), host_keys.get());
-        host_keys.update_data();
-
-        auto actual = ::std::is_heap(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-        // True only when n == 1
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(actual == (n == 1), "wrong result of is_heap_11");
-
-        actual = ::std::is_heap(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, first);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(actual == true, "wrong result of is_heap_12");
-
-        if (n <= 5)
-            return;
-
-        host_keys.retrieve_data();
-        ::std::make_heap(host_keys.get(), host_keys.get() + n / 2);
-        host_keys.update_data();
-
-        actual = ::std::is_heap(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(actual == false, "wrong result of is_heap_21");
-
-        auto end = first + n / 2;
-        actual = ::std::is_heap(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, end);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(actual == true, "wrong result of is_heap_22");
-
-        host_keys.retrieve_data();
-        ::std::make_heap(host_keys.get(), host_keys.get() + n);
-        host_keys.update_data();
-
-        actual = ::std::is_heap(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(actual == true, "wrong result of is_heap_3");
-    }
-};
-
-DEFINE_TEST(test_inplace_merge)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_inplace_merge)
-
-    template <typename Policy, typename Iterator, typename Size>
-    void
-    operator()(Policy&& exec, Iterator first, Iterator last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator>::value_type T;
-        auto value = T(0);
-
-        ::std::iota(host_keys.get(), host_keys.get() + n, value);
-
-        ::std::vector<T> exp(n);
-        ::std::iota(exp.begin(), exp.end(), value);
-
-        auto middle = ::std::stable_partition(host_keys.get(), host_keys.get() + n, [](const T& x) { return x % 2; });
-        host_keys.update_data();
-
-        ::std::inplace_merge(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, first + (middle - host_keys.get()), last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-
-        host_keys.retrieve_data();
-        for (size_t i = 0; i < n; ++i)
-        {
-            if (host_keys.get()[i] != exp[i])
-            {
-                ::std::cout << "Error: i = " << i << ", expected " << exp[i] << ", got " << host_keys.get()[i] << ::std::endl;
-            }
-            EXPECT_TRUE(host_keys.get()[i] == exp[i], "wrong effect from inplace_merge");
-        }
-    }
-};
-
-DEFINE_TEST(test_merge)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_merge)
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 /* last2 */, Iterator3 first3,
-               Iterator3 /* last3 */, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-        typedef typename ::std::iterator_traits<Iterator2>::value_type T2;
-        typedef typename ::std::iterator_traits<Iterator3>::value_type T3;
-
-        auto value = T1(0);
-        auto x = n > 1 ? n / 2 : n;
-        ::std::iota(host_keys.get(), host_keys.get() + n, value);
-        ::std::iota(host_vals.get(), host_vals.get() + n, T2(value));
-        update_data(host_keys, host_vals);
-
-        ::std::vector<T3> exp(2 * n);
-        auto exp1 = ::std::merge(host_keys.get(), host_keys.get() + n, host_vals.get(), host_vals.get() + x, exp.begin());
-        auto res1 = ::std::merge(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, first2 + x, first3);
-        TestDataTransfer<UDTKind::eRes, Size> host_res(*this, res1 - first3);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        // Special case, because we have more results then source data
-        host_res.retrieve_data();
-        auto host_first3 = host_res.get();
-#    if _ONEDPL_DEBUG_SYCL
-        for (size_t i = 0; i < res1 - first3; ++i)
-        {
-            if (host_first3[i] != exp[i])
-            {
-                ::std::cout << "Error: i = " << i << ", expected " << exp[i] << ", got " << host_first3[i] << ::std::endl;
-            }
-        }
-#    endif
-        EXPECT_TRUE(res1 - first3 == exp1 - exp.begin(), "wrong result from merge_1");
-        EXPECT_TRUE(::std::is_sorted(host_first3, host_first3 + (res1 - first3)), "wrong effect from merge_1");
-    }
-};
-
-DEFINE_TEST(test_sort)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_sort)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-
-        typedef typename ::std::iterator_traits<Iterator1>::value_type T1;
-
-        auto value = T1(333);
-        ::std::iota(host_keys.get(), host_keys.get() + n, value);
-        host_keys.update_data();
-
-        ::std::sort(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        {
-            host_keys.retrieve_data();
-            auto host_first1 = host_keys.get();
-#           if _ONEDPL_DEBUG_SYCL
-            for (int i = 0; i < n; ++i)
-            {
-                if (host_first1[i] != value + i)
-                {
-                    ::std::cout << "Error_1. i = " << i << ", expected = " << value + i << ", got = " << host_first1[i]
-                                << ::std::endl;
-                }
-            }
-#           endif
-            EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n), "wrong effect from sort_1");
-        }
-
-        ::std::sort(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, ::std::greater<T1>());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        auto host_first1 = host_keys.get();
-#       if _ONEDPL_DEBUG_SYCL
-        for (int i = 0; i < n; ++i)
-        {
-            if (host_first1[i] != value + n - 1 - i)
-            {
-                ::std::cout << "Error_2. i = " << i << ", expected = " << value + n - 1 - i
-                          << ", got = " << host_first1[i] << ::std::endl;
-            }
-        }
-#       endif
-        EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n, ::std::greater<T1>()), "wrong effect from sort_2");
     }
 };
 
@@ -2797,14 +1515,13 @@ DEFINE_TEST(test_stable_sort)
         host_keys.update_data();
 
         ::std::stable_sort(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         {
             host_keys.retrieve_data();
             auto host_first1 = host_keys.get();
 
-#           if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
             for (int i = 0; i < n; ++i)
             {
                 if (host_first1[i] != value + i)
@@ -2813,17 +1530,17 @@ DEFINE_TEST(test_stable_sort)
                               << ::std::endl;
                 }
             }
-#           endif
+#endif // _ONEDPL_DEBUG_SYCL
+
             EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n), "wrong effect from stable_sort_1");
         }
 
         ::std::stable_sort(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, ::std::greater<T1>());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         auto host_first1 = host_keys.get();
-#       if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         for (int i = 0; i < n; ++i)
         {
             if (host_first1[i] != value + n - 1 - i)
@@ -2832,7 +1549,8 @@ DEFINE_TEST(test_stable_sort)
                             << ", got = " << host_first1[i] << ::std::endl;
             }
         }
-#       endif
+#endif // _ONEDPL_DEBUG_SYCL
+
         EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n, ::std::greater<T1>()),
                     "wrong effect from stable_sort_3");
     }
@@ -2863,9 +1581,7 @@ DEFINE_TEST(test_partial_sort)
         {
             auto end1 = first1 + end_idx;
             ::std::partial_sort(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, end1, last1);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
 
             // Make sure that elements up to end are sorted and remaining elements are bigger
             // than the last sorted one.
@@ -2882,9 +1598,8 @@ DEFINE_TEST(test_partial_sort)
         if (end_idx > last1 - first1)
         {
             ::std::partial_sort(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, last1);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             host_keys.retrieve_data();
             auto host_first1 = host_keys.get();
             EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n), "wrong effect from partial_sort_2");
@@ -2920,9 +1635,8 @@ DEFINE_TEST(test_partial_sort_copy)
 
             auto last_sorted =
                 ::std::partial_sort_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, end2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             auto host_first1 = host_keys.get();
             auto host_first2 = host_vals.get();
 
@@ -2941,9 +1655,7 @@ DEFINE_TEST(test_partial_sort_copy)
         {
             auto last_sorted =
                 ::std::partial_sort_copy(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
 
             auto host_first1 = host_keys.get();
             auto host_first2 = host_vals.get();
@@ -2985,9 +1697,8 @@ DEFINE_TEST(test_find_end)
 
             // Empty subsequence
             auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_end_1");
 
             return;
@@ -3002,9 +1713,8 @@ DEFINE_TEST(test_find_end)
 
             // No subsequence
             auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, first2 + n / 2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_end_2");
 
             // Whole sequence is matched
@@ -3012,9 +1722,8 @@ DEFINE_TEST(test_find_end)
             host_keys.update_data();
 
             res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1, "Wrong effect from find_end_3");
 
             return;
@@ -3034,9 +1743,8 @@ DEFINE_TEST(test_find_end)
 
                 auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2,
                                            first2 + n / 5);
-#if _PSTL_SYCL_TEST_USM
-                exec.queue().wait_and_throw();
-#endif
+                wait_and_throw(exec);
+
                 EXPECT_TRUE(res == first1, "Wrong effect from find_end_4");
             }
 
@@ -3050,9 +1758,8 @@ DEFINE_TEST(test_find_end)
 
                 auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, first2,
                                            first2 + n / 5);
-#if _PSTL_SYCL_TEST_USM
-                exec.queue().wait_and_throw();
-#endif
+                wait_and_throw(exec);
+
                 EXPECT_TRUE(res == first1 + 2 * n / 5, "Wrong effect from find_end_5");
             }
 
@@ -3065,9 +1772,8 @@ DEFINE_TEST(test_find_end)
 
                 auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, first2,
                                            first2 + n / 5);
-#if _PSTL_SYCL_TEST_USM
-                exec.queue().wait_and_throw();
-#endif
+                wait_and_throw(exec);
+
                 EXPECT_TRUE(res == first1 + 4 * n / 5, "Wrong effect from find_end_6");
             }
         }
@@ -3104,9 +1810,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 1.1: S1 == S2 && len(S1) == len(S2)
         bool is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1,
                                                           last1, first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0, "wrong effect from lex_compare Test 1.1: S1 == S2 && len(S1) == len(S2)");
@@ -3114,9 +1819,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 1.2: S1 == S2 && len(S1) < len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1 - 1,
                                                    first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != 1)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 1" << ::std::endl;
         EXPECT_TRUE(is_less_res == 1, "wrong effect from lex_compare Test 1.2: S1 == S2 && len(S1) < len(S2)");
@@ -3124,9 +1828,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 1.3: S1 == S2 && len(S1) > len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2,
                                                    last2 - 1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0, "wrong effect from lex_compare Test 1.3: S1 == S2 && len(S1) > len(S2)");
@@ -3140,9 +1843,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 2.1: S1 < S2 (PRE-LAST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         bool is_less_exp = n > 1 ? 1 : 0;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -3152,9 +1854,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 2.2: S1 < S2 (PRE-LAST ELEMENT) && len(S1) > len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, first2,
                                                    last2 - 1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
         EXPECT_TRUE(is_less_res == is_less_exp,
@@ -3169,9 +1870,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 3.1: S1 > S2 (PRE-LAST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0,
@@ -3180,9 +1880,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 3.2: S1 > S2 (PRE-LAST ELEMENT) && len(S1) < len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 6>>(exec), first1, last1 - 1,
                                                    first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         is_less_exp = n > 1 ? 0 : 1;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -3196,9 +1895,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 4.1: S1 < S2 (FIRST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 7>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != 1)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 1" << ::std::endl;
         EXPECT_TRUE(is_less_res == 1, "wrong effect from lex_compare Test 4.1: S1 < S2 (FIRST) && len(S1) == len(S2)");
@@ -3206,9 +1904,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 4.2: S1 < S2 (FIRST ELEMENT) && len(S1) > len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 8>>(exec), first1, last1, first2,
                                                    last2 - 1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         is_less_exp = n > 1 ? 1 : 0;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -3222,9 +1919,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 5.1: S1 > S2 (FIRST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 9>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0, "wrong effect from lex_compare Test 5.1: S1 > S2 (FIRST) && len(S1) == len(S2)");
@@ -3232,9 +1928,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 5.2: S1 > S2 (FIRST ELEMENT) && len(S1) < len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 10>>(exec), first1, last1 - 1,
                                                    first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         is_less_exp = n > 1 ? 0 : 1;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -3263,9 +1958,8 @@ DEFINE_TEST(test_swap_ranges)
 
         Iterator2 actual_return = ::std::swap_ranges(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
 
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         bool check_return = (actual_return == last2);
         EXPECT_TRUE(check_return, "wrong result of swap_ranges");
         if (check_return)
@@ -3311,9 +2005,7 @@ DEFINE_TEST(test_nth_element)
         // invoke
         auto comp = ::std::less<T1>{};
         ::std::nth_element(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, middle1, last1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
 
         retrieve_data(host_keys, host_vals);
 
@@ -3338,34 +2030,6 @@ DEFINE_TEST(test_nth_element)
     }
 };
 
-DEFINE_TEST(test_reverse)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_reverse)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first, Iterator1 last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-        host_keys.retrieve_data();
-
-        using IteratorValyeType = typename ::std::iterator_traits<Iterator1>::value_type;
-
-        ::std::vector<IteratorValyeType> local_copy(n);
-        local_copy.assign(host_keys.get(), host_keys.get() + n);
-        ::std::reverse(local_copy.begin(), local_copy.end());
-
-        ::std::reverse(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        auto host_first1 = host_keys.get();
-        for (int i = 0; i < (last - first); ++i)
-            EXPECT_TRUE(local_copy[i] == host_first1[i], "wrong effect from reverse");
-    }
-};
-
 DEFINE_TEST(test_reverse_copy)
 {
     DEFINE_TEST_CONSTRUCTOR(test_reverse_copy)
@@ -3386,41 +2050,12 @@ DEFINE_TEST(test_reverse_copy)
         ::std::reverse(local_copy.begin(), local_copy.end());
 
         ::std::reverse_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, result_first);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         auto host_first2 = host_vals.get();
         for (int i = 0; i < n; ++i)
             EXPECT_TRUE(local_copy[i] == host_first2[i], "wrong effect from reverse_copy");
-    }
-};
-
-DEFINE_TEST(test_rotate)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_rotate)
-
-    template <typename Policy, typename Iterator1, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first, Iterator1 last, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, n);
-        host_keys.retrieve_data();
-
-        using IteratorValyeType = typename ::std::iterator_traits<Iterator1>::value_type;
-
-        ::std::vector<IteratorValyeType> local_copy(n);
-        local_copy.assign(host_keys.get(), host_keys.get() + n);
-        ::std::rotate(local_copy.begin(), local_copy.begin() + 1, local_copy.end());
-
-        ::std::rotate(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, first + 1, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_keys.retrieve_data();
-        auto host_first1 = host_keys.get();
-        for (int i = 0; i < (last - first); ++i)
-            EXPECT_TRUE(local_copy[i] == host_first1[i], "wrong effect from rotate");
     }
 };
 
@@ -3444,30 +2079,13 @@ DEFINE_TEST(test_rotate_copy)
         ::std::rotate(local_copy.begin(), local_copy.begin() + 1, local_copy.end());
 
         ::std::rotate_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, first + 1, last, result_first);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         for (int i = 0; i < n; ++i)
             EXPECT_TRUE(local_copy[i] == host_vals.get()[i], "wrong effect from rotate_copy");
     }
 };
-
-int a[] = {0, 0, 1, 1, 2, 6, 6, 9, 9};
-int b[] = {0, 1, 1, 6, 6, 9};
-int c[] = {0, 1, 6, 6, 6, 9, 9};
-int d[] = {7, 7, 7, 8};
-int e[] = {11, 11, 12, 16, 19};
-constexpr auto na = sizeof(a) / sizeof(a[0]);
-constexpr auto nb = sizeof(b) / sizeof(b[0]);
-constexpr auto nc = sizeof(c) / sizeof(c[0]);
-constexpr auto nd = sizeof(d) / sizeof(d[0]);
-
-template <typename Size>
-Size get_size(Size n)
-{
-    return n + na + nb + nc + nd;
-}
 
 DEFINE_TEST(test_includes)
 {
@@ -3490,9 +2108,8 @@ DEFINE_TEST(test_includes)
         host_vals.update_data(nb);
 
         auto result = ::std::includes(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result, "wrong effect from includes a, b");
 
         host_vals.retrieve_data();
@@ -3500,171 +2117,12 @@ DEFINE_TEST(test_includes)
         host_vals.update_data(nc);
 
         result = ::std::includes(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
+        wait_and_throw(exec);
 
         EXPECT_TRUE(!result, "wrong effect from includes a, c");
     }
 };
-
-DEFINE_TEST(test_set_intersection)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_set_intersection)
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 last2, Iterator3 first3,
-               Iterator3 last3, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, get_size(n));
-        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, get_size(n));
-        TestDataTransfer<UDTKind::eRes,  Size> host_res (*this, get_size(n));
-
-        //first test case
-        last1 = first1 + na;
-        last2 = first2 + nb;
-        ::std::copy(a, a + na, host_keys.get());
-        ::std::copy(b, b + nb, host_vals.get());
-        host_keys.update_data(na);
-        host_vals.update_data(nb);
-
-        last3 = ::std::set_intersection(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, last2,
-                                      first3);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        host_res.retrieve_data();
-        auto nres = last3 - first3;
-
-        EXPECT_TRUE(nres == 6, "wrong size of intersection of a, b");
-
-        auto result = ::std::includes(host_keys.get(), host_keys.get() + na, host_res.get(), host_res.get() + nres) &&
-                      ::std::includes(host_vals.get(), host_vals.get() + nb, host_res.get(), host_res.get() + nres);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        EXPECT_TRUE(result, "wrong effect from set_intersection a, b");
-
-        { //second test case
-
-            last2 = first2 + nd;
-            ::std::copy(a, a + na, host_keys.get());
-            ::std::copy(d, d + nd, host_vals.get());
-            host_keys.update_data(na);
-            host_vals.update_data(nb);
-
-            last3 = ::std::set_intersection(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2,
-                                          last2, first3);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif
-            auto nres = last3 - first3;
-            EXPECT_TRUE(nres == 0, "wrong size of intersection of a, d");
-        }
-    }
-};
-
-DEFINE_TEST(test_set_difference)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_set_difference)
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 last2, Iterator3 first3,
-               Iterator3 last3, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, get_size(n));
-        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, get_size(n));
-        TestDataTransfer<UDTKind::eRes,  Size> host_res (*this, get_size(n));
-
-        last1 = first1 + na;
-        last2 = first2 + nb;
-
-        ::std::copy(a, a + na, host_keys.get());
-        ::std::copy(b, b + nb, host_vals.get());
-        host_keys.update_data(na);
-        host_vals.update_data(nb);
-
-        last3 = ::std::set_difference(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, last2, first3);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        int res_expect[na];
-        host_res.retrieve_data();
-        auto nres_expect = ::std::set_difference(host_keys.get(), host_keys.get() + na, host_vals.get(), host_vals.get() + nb, res_expect) - res_expect;
-        EXPECT_EQ_N(host_res.get(), res_expect, nres_expect, "wrong effect from set_difference a, b");
-    }
-};
-
-DEFINE_TEST(test_set_union)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_set_union)
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 last2, Iterator3 first3,
-               Iterator3 last3, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, get_size(n));
-        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, get_size(n));
-        TestDataTransfer<UDTKind::eRes,  Size> host_res (*this, get_size(n));
-
-        last1 = first1 + na;
-        last2 = first2 + nb;
-
-        ::std::copy(a, a + na, host_keys.get());
-        ::std::copy(b, b + nb, host_vals.get());
-        host_keys.update_data(na);
-        host_vals.update_data(nb);
-
-        last3 = ::std::set_union(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, last2, first3);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        int res_expect[na + nb];
-        host_res.retrieve_data();
-        auto nres_expect =
-            ::std::set_union(host_keys.get(), host_keys.get() + na, host_vals.get(), host_vals.get() + nb, res_expect) - res_expect;
-        EXPECT_EQ_N(host_res.get(), res_expect, nres_expect, "wrong effect from set_union a, b");
-    }
-};
-
-DEFINE_TEST(test_set_symmetric_difference)
-{
-    DEFINE_TEST_CONSTRUCTOR(test_set_symmetric_difference)
-
-    template <typename Policy, typename Iterator1, typename Iterator2, typename Iterator3, typename Size>
-    void
-    operator()(Policy&& exec, Iterator1 first1, Iterator1 last1, Iterator2 first2, Iterator2 last2, Iterator3 first3,
-               Iterator3 last3, Size n)
-    {
-        TestDataTransfer<UDTKind::eKeys, Size> host_keys(*this, get_size(n));
-        TestDataTransfer<UDTKind::eVals, Size> host_vals(*this, get_size(n));
-        TestDataTransfer<UDTKind::eRes, Size>  host_res (*this, get_size(n));
-
-        last1 = first1 + na;
-        last2 = first2 + nb;
-
-        ::std::copy(a, a + na, host_keys.get());
-        ::std::copy(b, b + nb, host_vals.get());
-        host_keys.update_data(na);
-        host_vals.update_data(nb);
-
-        last3 = ::std::set_symmetric_difference(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1,
-                                                first2, last2, first3);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif
-        int res_expect[na + nb];
-        retrieve_data(host_keys, host_vals, host_res);
-        auto nres_expect = ::std::set_symmetric_difference(host_keys.get(), host_keys.get() + na, host_vals.get(),
-                                                           host_vals.get() + nb, res_expect) -
-                           res_expect;
-        EXPECT_EQ_N(host_res.get(), res_expect, nres_expect, "wrong effect from set_symmetric_difference a, b");
-    }
-};
-#endif
+#endif // TEST_DPCPP_BACKEND_PRESENT
 
 #if TEST_DPCPP_BACKEND_PRESENT
 template <sycl::usm::alloc alloc_type>
@@ -3672,93 +2130,6 @@ void
 test_usm_and_buffer()
 {
     using ValueType = ::std::int32_t;
-
-    // test1buffer
-    PRINT_DEBUG("test_for_each");
-    test1buffer<alloc_type, test_for_each<ValueType>>();
-    PRINT_DEBUG("test_for_each_n");
-    test1buffer<alloc_type, test_for_each_n<ValueType>>();
-    PRINT_DEBUG("test_replace");
-    test1buffer<alloc_type, test_replace<ValueType>>();
-    PRINT_DEBUG("test_replace_if");
-    test1buffer<alloc_type, test_replace_if<ValueType>>();
-    PRINT_DEBUG("test_fill");
-    test1buffer<alloc_type, test_fill<ValueType>>();
-    PRINT_DEBUG("test_fill_n");
-    test1buffer<alloc_type, test_fill_n<ValueType>>();
-    PRINT_DEBUG("test_generate");
-    test1buffer<alloc_type, test_generate<ValueType>>();
-    PRINT_DEBUG("test_generate_n");
-    test1buffer<alloc_type, test_generate_n<ValueType>>();
-    PRINT_DEBUG("test_reduce");
-    test1buffer<alloc_type, test_reduce<ValueType>>();
-    PRINT_DEBUG("test_transform_reduce_unary");
-    test1buffer<alloc_type, test_transform_reduce_unary<ValueType>>();
-    PRINT_DEBUG("test_any_all_none_of");
-    test1buffer<alloc_type, test_any_all_none_of<ValueType>>();
-    PRINT_DEBUG("test_is_sorted");
-    test1buffer<alloc_type, test_is_sorted<ValueType>>();
-    PRINT_DEBUG("test_count");
-    test1buffer<alloc_type, test_count<ValueType>>();
-    PRINT_DEBUG("test_count_if");
-    test1buffer<alloc_type, test_count_if<ValueType>>();
-    PRINT_DEBUG("test_is_partitioned");
-    test1buffer<alloc_type, test_is_partitioned<ValueType>>();
-    PRINT_DEBUG("test_sort");
-    test1buffer<alloc_type, test_sort<ValueType>>();
-    PRINT_DEBUG("test_min_element");
-    test1buffer<alloc_type, test_min_element<ValueType>>();
-    PRINT_DEBUG("test_max_element");
-    test1buffer<alloc_type, test_max_element<ValueType>>();
-    PRINT_DEBUG("test_minmax_element");
-    test1buffer<alloc_type, test_minmax_element<ValueType>>();
-    PRINT_DEBUG("test_inplace_merge");
-    test1buffer<alloc_type, test_inplace_merge<ValueType>>();
-    PRINT_DEBUG("test_reverse");
-    test1buffer<alloc_type, test_reverse<ValueType>>();
-    PRINT_DEBUG("test_rotate");
-    test1buffer<alloc_type, test_rotate<ValueType>>();
-    PRINT_DEBUG("test_partition");
-    test1buffer<alloc_type, test_partition<ValueType>>();
-    PRINT_DEBUG("test_is_heap");
-    test1buffer<alloc_type, test_is_heap<ValueType>>();
-    PRINT_DEBUG("test_find_if");
-    test1buffer<alloc_type, test_find_if<ValueType>>();
-    PRINT_DEBUG("test_adjacent_find");
-    test1buffer<alloc_type, test_adjacent_find<ValueType>>();
-    PRINT_DEBUG("test_is_sorted_until");
-    test1buffer<alloc_type, test_is_sorted_until<ValueType>>();
-    PRINT_DEBUG("test_search_n");
-    test1buffer<alloc_type, test_search_n<ValueType>>();
-    PRINT_DEBUG("test_remove");
-    test1buffer<alloc_type, test_remove<ValueType>>();
-    PRINT_DEBUG("test_remove_if");
-    test1buffer<alloc_type, test_remove_if<ValueType>>();
-    PRINT_DEBUG("test_stable_sort");
-    test1buffer<alloc_type, test_stable_sort<ValueType>>();
-    PRINT_DEBUG("test_unique");
-    test1buffer<alloc_type, test_unique<ValueType>>();
-    PRINT_DEBUG("test_is_heap_until");
-    test1buffer<alloc_type, test_is_heap_until<ValueType>>();
-    PRINT_DEBUG("test_uninitialized_fill");
-    test1buffer<alloc_type, test_uninitialized_fill<ValueType>>();
-    PRINT_DEBUG("test_uninitialized_fill_n");
-    test1buffer<alloc_type, test_uninitialized_fill_n<ValueType>>();
-    PRINT_DEBUG("test_uninitialized_default_construct");
-    test1buffer<alloc_type, test_uninitialized_default_construct<SyclTypeWrapper<ValueType>>>();
-    PRINT_DEBUG("test_uninitialized_default_construct_n");
-    test1buffer<alloc_type, test_uninitialized_default_construct_n<SyclTypeWrapper<ValueType>>>();
-    PRINT_DEBUG("test_uninitialized_value_construct");
-    test1buffer<alloc_type, test_uninitialized_value_construct<ValueType>>();
-    PRINT_DEBUG("test_uninitialized_value_construct_n");
-    test1buffer<alloc_type, test_uninitialized_value_construct_n<ValueType>>();
-    print_debug("test_is_heap");
-    test1buffer<alloc_type, test_is_heap<ValueType>>();
-    PRINT_DEBUG("test_destroy");
-    test1buffer<alloc_type, test_destroy<SyclTypeWrapper<ValueType>>>();
-    PRINT_DEBUG("test_destroy_n");
-    test1buffer<alloc_type, test_destroy_n<SyclTypeWrapper<ValueType>>>();
-    test1buffer<alloc_type, test_destroy_n<ValueType>>();
 
     //test2buffers
     PRINT_DEBUG("test_nth_element");
@@ -3821,22 +2192,8 @@ test_usm_and_buffer()
     test2buffers<alloc_type, test_find_first_of<ValueType>>();
     PRINT_DEBUG("test_includes");
     test2buffers<alloc_type, test_includes<ValueType>>();
-
-    //test3buffers
-    PRINT_DEBUG("test_partition_copy");
-    test3buffers<alloc_type, test_partition_copy<ValueType>>();
-    PRINT_DEBUG("test_set_symmetric_difference");
-    test3buffers<alloc_type, test_set_symmetric_difference<ValueType>>();
-    PRINT_DEBUG("test_set_union");
-    test3buffers<alloc_type, test_set_union<ValueType>>();
-    PRINT_DEBUG("test_set_difference");
-    test3buffers<alloc_type, test_set_difference<ValueType>>();
-    PRINT_DEBUG("test_set_intersection");
-    test3buffers<alloc_type, test_set_intersection<ValueType>>();
-    PRINT_DEBUG("test_merge");
-    test3buffers<alloc_type, test_merge<ValueType>>(2);
 }
-#endif
+#endif // TEST_DPCPP_BACKEND_PRESENT
 
 std::int32_t
 main()
